requirements.txt
azure-ai-textanalytics==5.3.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.2.2
tabulate==0.9.0
openai>=1.11.0
streamlit>=1.40.0
azure-cosmos


